PIM-3595: move MongoDB ODM types in doctrine folder
